+++ CHANGES.rst
@@ -19,6 +19,8 @@ Features
   server in read-only mode.
 - Handles connecting to Zookeeper 3.4+ read-only servers.
 - Added support for Python 3.2, 3.3 and PyPy (only for the threading handler).
+- Added a new `retry_max_delay` argument to the client and by default limit the
+  retry delay to at most an hour regardless of exponential backoff settings.
 
 Bug Handling
 ************